+++ includes/functions/parser.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parser.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/08 20:41:18 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/15 17:03:44 by jvacaris         ###   ########.fr       */
+/*   Updated: 2022/03/15 23:02:02 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,9 @@ t_item		*line2sphere(char **line);
 t_item		*line2plane(char **line);
 t_item		*line2cylinder(char **line);
 
-t_itemlist	**get_items(char *argv);
+void		get_items(t_itemlist **list,char *argv);
 t_corners	set_corners(t_item item);
 t_coords	**set_cam_vectors(t_corners corn);
+
 
 #endif