+++ src/parser/fill_struct_by_type.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   fill_struct_by_type.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/08 18:11:58 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/09 17:47:13 by jvacaris         ###   ########.fr       */
+/*   Updated: 2022/03/15 22:38:36 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,36 +15,30 @@
 t_item	*line2alight(char **line)
 {
 	t_item	*item;
-	char	**split_commas;
+	char	**rgb;
 
 	item = malloc(sizeof(t_item));
 	item->type = ALIGHT;
 	item->brightness = ft_atof(line[1]);
-	split_commas = ft_split(line[2], ',');
-	item->color.r = ft_atof(split_commas[0]);
-	item->color.g = ft_atof(split_commas[1]);
-	item->color.b = ft_atof(split_commas[2]);
-	megafree(&split_commas);
+	rgb = ft_split(line[2], ',');
+	item->color = set_color(rgb);
+	megafree(&rgb);
 	return (item);
 }
 
 t_item	*line2camera(char **line)
 {
 	t_item	*item;
-	char	**split_commas;
+	char	**coords;
 
 	item = malloc(sizeof(t_item));
 	item->type = CAMERA;
-	split_commas = ft_split(line[1], ',');
-	item->loc.x = ft_atof(split_commas[0]);
-	item->loc.y = ft_atof(split_commas[1]);
-	item->loc.z = ft_atof(split_commas[2]);
-	megafree(&split_commas);
-	split_commas = ft_split(line[2], ',');
-	item->orient.x = ft_atof(split_commas[0]);
-	item->orient.y = ft_atof(split_commas[1]);
-	item->orient.z = ft_atof(split_commas[2]);
-	megafree(&split_commas);
+	coords = ft_split(line[1], ',');
+	item->loc = set_coords(coords);
+	megafree(&coords);
+	coords = ft_split(line[2], ',');
+	item->orient = set_coords(coords);
+	megafree(&coords);
 	item->fov = ft_atof(line[3]);
 	return (item);
 }
@@ -52,15 +46,13 @@ t_item	*line2camera(char **line)
 t_item	*line2light(char **line)
 {
 	t_item	*item;
-	char	**split_commas;
+	char	**coords;
 
 	item = malloc(sizeof(t_item));
 	item->type = LIGHT;
-	split_commas = ft_split(line[1], ',');
-	item->loc.x = ft_atof(split_commas[0]);
-	item->loc.y = ft_atof(split_commas[1]);
-	item->loc.z = ft_atof(split_commas[2]);
-	megafree(&split_commas);
+	coords = ft_split(line[1], ',');
+	item->loc = set_coords(coords);
+	megafree(&coords);
 	item->brightness = ft_atof(line[2]);
 	return (item);
 }
@@ -68,72 +60,57 @@ t_item	*line2light(char **line)
 t_item	*line2sphere(char **line)
 {
 	t_item	*item;
-	char	**split_commas;
+	char	**split;
 
 	item = malloc(sizeof(t_item));
 	item->type = SPHERE;
-	split_commas = ft_split(line[1], ',');
-	item->loc.x = ft_atof(split_commas[0]);
-	item->loc.y = ft_atof(split_commas[1]);
-	item->loc.z = ft_atof(split_commas[2]);
-	megafree(&split_commas);
+	split = ft_split(line[1], ',');
+	item->loc = set_coords(split);
+	megafree(&split);
 	item->diameter = ft_atof(line[2]);
-	split_commas = ft_split(line[3], ',');
-	item->color.r = ft_atof(split_commas[0]);
-	item->color.g = ft_atof(split_commas[1]);
-	item->color.b = ft_atof(split_commas[2]);
-	megafree(&split_commas);
+	split = ft_split(line[3], ',');
+	item->color = set_color(split);
+	megafree(&split);
 	return (item);
 }
 
 t_item	*line2plane(char **line)
 {
 	t_item	*item;
-	char	**split_commas;
+	char	**split;
 
 	item = malloc(sizeof(t_item));
 	item->type = PLANE;
-	split_commas = ft_split(line[1], ',');
-	item->loc.x = ft_atof(split_commas[0]);
-	item->loc.y = ft_atof(split_commas[1]);
-	item->loc.z = ft_atof(split_commas[2]);
-	megafree(&split_commas);
-	split_commas = ft_split(line[2], ',');
-	item->orient.x = ft_atof(split_commas[0]);
-	item->orient.y = ft_atof(split_commas[1]);
-	item->orient.z = ft_atof(split_commas[2]);
-	megafree(&split_commas);
-	split_commas = ft_split(line[3], ',');
-	item->color.r = ft_atof(split_commas[0]);
-	item->color.g = ft_atof(split_commas[1]);
-	item->color.b = ft_atof(split_commas[2]);
-	megafree(&split_commas);
+	split = ft_split(line[1], ',');
+	item->loc = set_coords(split);
+	megafree(&split);
+	split = ft_split(line[2], ',');
+	item->orient = set_coords(split);
+	megafree(&split);
+	split = ft_split(line[3], ',');
+	item->color = set_color(split);
+	megafree(&split);
 	return (item);
 }
 
 t_item	*line2cylinder(char **line)
 {
 	t_item	*item;
-	char	**split_commas;
+	char	**split;
 
 	item = malloc(sizeof(t_item));
 	item->type = CYLINDER;
-	split_commas = ft_split(line[1], ',');
-	item->loc.x = ft_atof(split_commas[0]);
-	item->loc.y = ft_atof(split_commas[1]);
-	item->loc.z = ft_atof(split_commas[2]);
-	megafree(&split_commas);
-	split_commas = ft_split(line[2], ',');
-	item->orient.x = ft_atof(split_commas[0]);
-	item->orient.y = ft_atof(split_commas[1]);
-	item->orient.z = ft_atof(split_commas[2]);
-	megafree(&split_commas);
+	split = ft_split(line[1], ',');
+	item->loc = set_coords(split);
+	megafree(&split);
+	split = ft_split(line[2], ',');
+	item->orient = set_coords(split);
+	megafree(&split);
 	ft_atof(line[3]);
 	ft_atof(line[4]);
-	split_commas = ft_split(line[5], ',');
-	item->color.r = ft_atof(split_commas[0]);
-	item->color.g = ft_atof(split_commas[1]);
-	item->color.b = ft_atof(split_commas[2]);
-	megafree(&split_commas);
+	split = ft_split(line[5], ',');
+	item->color = set_color(split);
+	megafree(&split);
 	return (item);
 }
+