+++ CHANGELOG.md
@@ -5,10 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2022-03-15 - Edu
+### Added
+- Functions to print list_rt **lst_rt_print** and filter by type **get_items_by_type**
+### Changed
+- Some documentation on utils
+### Fixed
+- Some leaks at parsing list
+
 ## 2022-03-13 - Edu
 ### Added
 - Add CHANGELOG.md, README.md, Makefile
-
 ### Changed
 - Structure files in folders.